Switch to source importing and allow additional imports.
Go 1.5 introduced the 'source' compiler. See
https://github.com/golang/go/issues/11415 for details.

This avoids the problem of using out of date information when generating
code. It also means we don't have to have to run `go install` prior to
generating.

I also added the ability to specify additional imports since we use
codegen to generate files that depend on types outside testify.